+++ upload/lib/Cms/Exception/Ajax/AjaxException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Cms\Exception\Ajax;
+
+
+class AjaxException extends \Exception
+{
+    
+}